Add files via upload
This Python project allows users to generate a Spotify playlist of the top 100 songs on the Billboard chart for a given date. It uses web scraping to retrieve the song titles and artists from the Billboard website, and then uses the Spotipy API to create the playlist on the user's Spotify account.